refactor: consolidate workflow handlers into handlers.py
Merge workflow_handlers.py and workflow_progress.py into handlers.py:
- All workflow progress tracking now in single file
- Renamed functions for clarity (progress_handler -> on_workflow_progress)
- Removed duplicate _emit_delta (identical to _emit_state_delta)
- Updated test imports

Files removed:
- workflow_handlers.py
- workflow_progress.py
